fix: Update routing to use proper page URLs
- Move disclosure viewer from / to /disclosures
- Add home page that redirects to /dashboard
- Update navigation links to point to /disclosures
- Ensure all pages have descriptive URLs matching their purpose
